# Refactor migrations to only marshall yaml twice