@jesseduffield 2025-02-25
21ms on each startup is indeed something, albeit a small improvement. But the complexity cost is high and there's a lot more chance of errors now that various places need to report on whether a change occurred, and any one of those might report incorrectly.

What's it like if we just do a deep comparison at the end before marshalling back? I suspect that would give us roughly the same speed up, without the extra complexity.