@ChrisMcD1 2025-02-25
Pushed a new version that does the unmarshalling twice. That's the only way I know to do a deep copy! Alternatively happily welcome.

My computer was seemingly feeling snappier today, so the test results were as follows:

Existing code
```
goos: linux
goarch: amd64
pkg: github.com/jesseduffield/lazygit/pkg/config
cpu: Intel(R) Core(TM) i7-8705G CPU @ 3.10GHz
BenchmarkMigrationOnLargeConfiguration-8             246           4828826 ns/op         1501479 B/op      19316 allocs/op
BenchmarkMigrationOnLargeConfiguration-8             244           4767374 ns/op         1501489 B/op      19316 allocs/op
BenchmarkMigrationOnLargeConfiguration-8             247           4888204 ns/op         1501483 B/op      19316 allocs/op
BenchmarkMigrationOnLargeConfiguration-8             241           4746350 ns/op         1501507 B/op      19316 allocs/op
BenchmarkMigrationOnLargeConfiguration-8             247           4737518 ns/op         1501484 B/op      19316 allocs/op
BenchmarkMigrationOnLargeConfiguration-8             249           4813931 ns/op         1501491 B/op      19316 allocs/op
BenchmarkMigrationOnLargeConfiguration-8             244           4749415 ns/op         1501504 B/op      19316 allocs/op
BenchmarkMigrationOnLargeConfiguration-8             246           4883387 ns/op         1501523 B/op      19316 allocs/op
BenchmarkMigrationOnLargeConfiguration-8             255           4785092 ns/op         1501493 B/op      19316 allocs/op
BenchmarkMigrationOnLargeConfiguration-8             242           4764061 ns/op         1501510 B/op      19317 allocs/op
```

Previous Implementation
```
goos: linux
goarch: amd64
pkg: github.com/jesseduffield/lazygit/pkg/config
cpu: Intel(R) Core(TM) i7-8705G CPU @ 3.10GHz
BenchmarkMigrationOnLargeConfiguration-8             913           1147582 ns/op          265811 B/op       3928 allocs/op
BenchmarkMigrationOnLargeConfiguration-8            1316            857591 ns/op          265801 B/op       3928 allocs/op
BenchmarkMigrationOnLargeConfiguration-8            1274            885818 ns/op          265792 B/op       3928 allocs/op
BenchmarkMigrationOnLargeConfiguration-8            1282            920565 ns/op          265789 B/op       3928 allocs/op
BenchmarkMigrationOnLargeConfiguration-8            1347            945753 ns/op          265794 B/op       3928 allocs/op
BenchmarkMigrationOnLargeConfiguration-8            1353            861908 ns/op          265794 B/op       3928 allocs/op
BenchmarkMigrationOnLargeConfiguration-8            1286            852052 ns/op          265796 B/op       3928 allocs/op
BenchmarkMigrationOnLargeConfiguration-8            1401            845480 ns/op          265797 B/op       3928 allocs/op
BenchmarkMigrationOnLargeConfiguration-8            1402            822486 ns/op          265795 B/op       3928 allocs/op
BenchmarkMigrationOnLargeConfiguration-8            1461            820271 ns/op          265799 B/op       3928 allocs/op
```

New implementation
```
goos: linux
goarch: amd64
pkg: github.com/jesseduffield/lazygit/pkg/config
cpu: Intel(R) Core(TM) i7-8705G CPU @ 3.10GHz
BenchmarkMigrationOnLargeConfiguration-8             530           2237977 ns/op          597309 B/op       7034 allocs/op
BenchmarkMigrationOnLargeConfiguration-8             560           2119147 ns/op          597395 B/op       7035 allocs/op
BenchmarkMigrationOnLargeConfiguration-8             559           2224450 ns/op          597347 B/op       7035 allocs/op
BenchmarkMigrationOnLargeConfiguration-8             570           2302126 ns/op          597431 B/op       7035 allocs/op
BenchmarkMigrationOnLargeConfiguration-8             601           1952700 ns/op          597492 B/op       7035 allocs/op
BenchmarkMigrationOnLargeConfiguration-8             607           1969020 ns/op          597366 B/op       7035 allocs/op
BenchmarkMigrationOnLargeConfiguration-8             565           2019408 ns/op          597369 B/op       7035 allocs/op
BenchmarkMigrationOnLargeConfiguration-8             613           1949337 ns/op          597403 B/op       7035 allocs/op
BenchmarkMigrationOnLargeConfiguration-8             592           1997986 ns/op          597363 B/op       7035 allocs/op
BenchmarkMigrationOnLargeConfiguration-8             603           1955303 ns/op          597433 B/op       7035 allocs/op
```

I'm still happy with new implementation though, since it'll make future migrations cost basically nothing.