@stefanhaller 2025-02-26
> Pushed a new version that does the unmarshalling twice. That's the only way I know to do a deep copy! Alternatively happily welcome.

Might be worth looking into https://github.com/brunoga/deep instead? I have no experience with it myself, just found it with a cursory google search.