@jesseduffield 2025-02-26
Something that just occurred to me: what if we just had some check at the start to see if the current yaml satisfies the schema? if so, we don't need to do any migrating. Then it won't matter how long it takes to do the migration because it will only need to run occasionally (after a user updates lazygit)