@ChrisMcD1 2025-02-27
@jesseduffield That worked quite smoothly! I spun it off into its own PR, because it ends up being a totally separate section of code: https://github.com/jesseduffield/lazygit/pull/4334

I've removed the benchmark commit from this branch, since it is no longer a relevant thing to be tracking.

I still think this refactor stands on its own merits. It feels more elegant for me for the individual migrations to be expressed in terms of acting on the root `yaml.Node`, instead of on a `[]byte`, and each of them no longer need to have error handling for invalid YAML marshalling or unmarshalling. See https://github.com/ChrisMcD1/lazygit/pull/1 for what the N+1 of a new migration looks like if the interface is in terms of a yaml node. (That PR will be turned into a PR on this repo once we decide the fate of this PR, as it is on a stacked branch)