@stefanhaller 2025-02-27
> Something that just occurred to me: what if we just had some check at the start to see if the current yaml satisfies the schema? if so, we don't need to do any migrating. Then it won't matter how long it takes to do the migration because it will only need to run occasionally (after a user updates lazygit)

I'm against this approach, because I'm not sure we can really trust the schema validation enough for this. Here's one example: the schema declares the context of a custom command as a string, because we can't do any better; conceptually it's a bunch of enums separated by commas. Now suppose we decide to rename one of the enum entries (e.g. we find a better name for "commitFiles"). So we write a migrator that splits the context by comma, and renames every occurrence of "commitFiles" to that better name. Schema validation doesn't catch this, because it's still a string before and after.

Maybe this is a contrived example, but it does illustrate the problem. I'd much rather we use the actual migrations that we perform to tell us whether anything happened.

I still find the deepCopy/deepEquals approach the most promising.