@stefanhaller 2025-02-27
> @jesseduffield That worked quite smoothly! I spun it off into its own PR, because it ends up being a totally separate section of code: #4334

@ChrisMcD1 I'm confused, can you explain more what you're doing in that PR? It doesn't seem to have anything to do with Jesse's suggestion as far as I can see.