@jesseduffield 2025-02-27
Good points @stefanhaller